cmake: offer CMake modules with no hard-coded path
Hardcoding absolute install paths in the CMake module files totally
breaks the cross-compilation: it makes impossible to be used in a
sysroot because the hardcoded paths are the ones of the target, whereas
they should point to the sysroot location. This tootally breaks the
CMake FIND_ROOT_PATH mechanism designed to handle cross-compilation.

This patch introduces a new CMake variable: NO_ABSOLUTE_PATH_IN_MODULE,
which, when set on the cmake command-line, will enable the installation
of the CMake module files using the CMake primitives find_library and
find_path to set the libraries and headers' locations, instead of
hard-coding them.

Thus, this patch adds non-templated CMake module files for octomap,
dynamicEDT3D and octovis for the non hard-coded path version of the
modules.

In any way, only the installed CMake module files are changed. The
templated ones are still needed for combined package build.

The non hard-coded path version of the modules are only installed when
NO_ABSOLUTE_PATH_IN_MODULE is set; by default the existing behavior
remains unchanged to keep honoring the behavior described in [1].

[1] https://github.com/OctoMap/octomap/pull/57#issuecomment-38161187