Adding option to choose if octovis&dynamicEDT3D are built
As suggested in an issue I added the option to the main cmakelists :
BUILD_OCTOVIS_SUBPROJECT.
If it's ON (default) everything remains the same, if it's turned off the
octomap project can be built even if no QGLViewer and/or OpenGL are
found on the system.
It also can be used when octomap is used as an external project in a
cmake-project, the option can be passed in the cmake_args.