Enable maximum tree depth of 32 (..)
The depth of an octree was limited to be exactly 16, so that the space
that could be mapped was limited in size by 2^16*resolution meter, eg
to approx. 6000m at a resolution of 0.1m. To enable larger tree dephts
some OcTreeKey related stuff needed to be changed:

a) Change all short int to unsigned int

And, at some places where such a change occured, unsigned int to long int.
Also changed some constants for 2^16 to 2^32. Utlimately, std::uint8_t et
al should be used.

b) Add a tricky hash function for 32bit based OcTreeKey

OcTreeKey consists of 3 32bit unsigned integers, but a hash function that
is to be used with the C++ std library needs to return a 64bit std::size_t
as a hash value. So we can't generate unique hash keys for every node
anymore. The solution is to only use the lower 16bit of every OcTreeKey-
component, so the resulting 3x16=48bit perfectly fit into std::size_t. This
is only possible as the hash function is never used for all the tree, but
only for the current point cloud. Points will only differ in their last
bits, probably, so the hash values generated for all points of one point
cloud should be unique again.