Make octomap-config.cmake.in cross-compilation proof
The file octomap-config.cmake.in may be used not only for
on-target compilation, but also in a cross-compilation environment
where the file is placed not under ${CMAKE_INSTALL_PREFIX}, but
under ${SOME_SYSROOT}${CMAKE_INSTALL_PREFIX} like in case of Yocto.

In order to accommodate both cases (when a dependent lib is
cross-compiled first, then gets recompiled on the target device)
it's more reliable not use absolute pathes to the libs in the
config included by dependent source packages. Instead use
CMake's find_library() command like it's been done for packages
like urdfdom.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>