improved insertPointCloudRays
An option to specify if the endpoint is occupied was added. This can be used when an obstacle is farher than the range sensor measuring limit. The ray can be used to update free nodes while not updating any node as occupied. The function is now overloaded and still accpets the old version.