Remove unused ament_cmake dependency.
REP-136 recommends that third-party (non-ROS) packages add a runtime
dependency on catkin so that the setup scripts provided by the catkin
package which set up a ROS environment are installed along with that
package. In ROS 2, those scripts are provided by the ros_workspace
package and bloom transparently injects a dependency on ros_workspace
into every released package.

It would be possible to use ament_cmake (or more likely the individual
ament_cmake_core package to optionally provide the package.xml and
ament resource index registration but since what you have is working
fine I think just removing the unecessary dependency is a good next
step. As mentioned in
https://github.com/flexible-collision-library/fcl/pull/536#issuecomment-845533649
we haven't produced a definitive recommendation for third party packages
in ROS 2 but I've added it to a meeting agenda and we'll document the
conclusions of that discussion and share them as well.